Allow different size radiuses
**This Commit**
Removes the `CIRCLE_RADIUS` constant in favor of per-circle radiuses.

**Note**
TBD on whether or not there should be a separate vec of radiuses. I
don't think so since I think the vector we want to optimize is "the one
used for collision detection" and for that we need the origins _and_ the
radiuses.